requirements.txt
torch
torchvision
opencv-python
mediapipe
pyserial
Pillow
tqdm
